Fix genre context menu having no playlists on "Add to playlist" submenu
This happened only from time to time, after switching to other tabs or
pages. The menu would appear only with the "New playlist" entry. It
would fix itself after switching to other tabs (not always).

The issue was the playlist list passed to the genre list composable
wasn't using a reactive type, so it wouldn't trigger a recomposition if
it changed. I guess there's some race condition that make it work most
of the time.